Delegating runner behaviour to runner class
