+++ lib/mumukit/server.rb
@@ -3,6 +3,4 @@ module Mumukit
   end
 end
 
-require_relative './server/response_builder'
-require_relative './server/test_pipeline'
 require_relative './server/test_server'
